// ui/packages/src/api/generated/filesys/schemas/initTransferUploadCommand.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Refactored free-fs with a subject-match-java style layered architecture.
 * OpenAPI spec version: v1.0.0
 */

export interface InitTransferUploadCommand {
  /** No comments found. */
  chunkSize: number;
  /** No comments found. */
  fileName: string;
  /** No comments found. */
  fileSize: number;
  /** No comments found. */
  mimeType: string;
  /** No comments found. */
  overwriteExisting?: boolean;
  /** No comments found. */
  parentId?: string;
  /** No comments found. */
  storageSettingId?: string;
  /** No comments found. */
  totalChunks: number;
  /** No comments found. */
  userId?: string;
}